refactor: extract timer bootverify checks
