+++ internal/runtime/bootverify/checks.go
@@ -13,7 +13,6 @@ import (
 
 	runtimecontracts "swarm/internal/runtime/contracts"
 	"swarm/internal/runtime/core/eventidentity"
-	"swarm/internal/runtime/core/timeridentity"
 	runtimecredentials "swarm/internal/runtime/credentials"
 	runtimemcp "swarm/internal/runtime/mcp"
 	runtimepipeline "swarm/internal/runtime/pipeline"
@@ -240,7 +239,6 @@ func checkAgentPermissionValidation(c *checkerContext) []Finding {
 func checkPhantomProduces(c *checkerContext) []Finding            { return c.phantomProduces() }
 func checkSingleNodePerEvent(c *checkerContext) []Finding         { return c.singleNodePerEvent() }
 func checkPlatformMetadataValidation(c *checkerContext) []Finding { return c.platformMetadata() }
-func checkTimerValidation(c *checkerContext) []Finding            { return c.timerValidation() }
 func checkGateSchemaValidation(c *checkerContext) []Finding       { return c.gateSchemaValidation() }
 func checkDeprecatedContractAlias(c *checkerContext) []Finding    { return c.deprecatedAliases() }
 
@@ -311,113 +309,6 @@ func (c *checkerContext) platformMetadata() []Finding {
 		})
 	}
 	return c.platformMetaFindings
-}
-
-func (c *checkerContext) timerValidation() []Finding {
-	if c.timerLoaded {
-		return c.timerFindings
-	}
-	c.timerLoaded = true
-	for _, timer := range c.source.WorkflowTimers() {
-		owner := strings.TrimSpace(timer.Owner)
-		if owner == "" {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "error",
-				Message:  fmt.Sprintf("timer %s missing owner", timer.ID),
-				Location: strings.TrimSpace(timer.ID),
-			})
-			continue
-		}
-		if owner != "runtime" {
-			if _, systemNode := c.source.NodeEntries()[owner]; !systemNode {
-				if !participantExistsLocal(c.source, owner) {
-					c.timerFindings = append(c.timerFindings, Finding{
-						CheckID:  "timer_validation",
-						Severity: "error",
-						Message:  fmt.Sprintf("timer %s owner %s missing from participants", timer.ID, owner),
-						Location: strings.TrimSpace(timer.ID),
-					})
-				}
-			}
-		}
-		if !eventExists(c.source, strings.TrimSpace(timer.Event)) {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "error",
-				Message:  fmt.Sprintf("timer %s event %s missing from event catalog", timer.ID, timer.Event),
-				Location: strings.TrimSpace(timer.ID),
-			})
-		}
-		startTrigger, err := timeridentity.ParseStartTrigger(timer.StartOn)
-		if err != nil {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "error",
-				Message:  fmt.Sprintf("timer %s start_on %q is invalid: %v", timer.ID, timer.StartOn, err),
-				Location: strings.TrimSpace(timer.ID),
-			})
-		} else {
-			c.validateTimerTrigger(timer, "start_on", startTrigger)
-		}
-		cancelTrigger, err := timeridentity.ParseCancelTrigger(timer.CancelOn)
-		if err != nil {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "error",
-				Message:  fmt.Sprintf("timer %s cancel_on %q is invalid: %v", timer.ID, timer.CancelOn, err),
-				Location: strings.TrimSpace(timer.ID),
-			})
-		} else {
-			c.validateTimerTrigger(timer, "cancel_on", cancelTrigger)
-		}
-	}
-	return c.timerFindings
-}
-
-func (c *checkerContext) validateTimerTrigger(timer runtimecontracts.WorkflowTimerContract, field string, trigger timeridentity.Trigger) {
-	if !trigger.Valid() {
-		return
-	}
-	switch trigger.Kind {
-	case timeridentity.TriggerKindState:
-		if !containsString(flowStatesForTimer(c.source, timer.FlowID), trigger.Name) {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "error",
-				Message:  fmt.Sprintf("timer %s %s references unknown state %s", timer.ID, field, trigger.Name),
-				Location: strings.TrimSpace(timer.ID),
-			})
-		}
-	case timeridentity.TriggerKindEvent:
-		if !eventExists(c.source, trigger.Name) {
-			c.timerFindings = append(c.timerFindings, Finding{
-				CheckID:  "timer_validation",
-				Severity: "warning",
-				Message:  fmt.Sprintf("timer %s %s references unknown event %s", timer.ID, field, trigger.Name),
-				Location: strings.TrimSpace(timer.ID),
-			})
-		}
-	}
-}
-
-func flowStatesForTimer(source semanticview.Source, flowID string) []string {
-	flowID = strings.TrimSpace(flowID)
-	if source == nil {
-		return nil
-	}
-	if flowID != "" {
-		return source.FlowStates(flowID)
-	}
-	stages := source.WorkflowStages()
-	out := make([]string, 0, len(stages))
-	for _, stage := range stages {
-		name := strings.TrimSpace(stage.ID)
-		if name != "" {
-			out = append(out, name)
-		}
-	}
-	return out
 }
 
 func (c *checkerContext) gateSchemaValidation() []Finding {
@@ -2538,25 +2429,6 @@ func eventExists(source semanticview.Source, eventType string) bool {
 	}
 	for candidate := range source.EventEntries() {
 		if routeMatchesLocal(eventType, strings.TrimSpace(candidate)) {
-			return true
-		}
-	}
-	return false
-}
-
-func participantExistsLocal(source semanticview.Source, participant string) bool {
-	participant = strings.TrimSpace(participant)
-	if participant == "" || source == nil {
-		return false
-	}
-	if participant == "runtime" || participant == "human" {
-		return true
-	}
-	if _, ok := source.NodeEntries()[participant]; ok {
-		return true
-	}
-	for _, agent := range source.AgentEntries() {
-		if strings.TrimSpace(agent.ID) == participant || strings.TrimSpace(agent.Role) == participant {
 			return true
 		}
 	}
